moved auth code flow and pkce content to BuildingOIDC content file, removed unused tsx files
