--- src/components/oauthDocs/AuthCodeFlowContent.tsx
@@ -1,32 +1,0 @@
-import * as React from 'react';
-import { useSelector } from 'react-redux';
-import { isApiDeactivated } from '../../apiDefs/deprecated';
-import { getAllOauthApis, lookupApiByFragment } from '../../apiDefs/query';
-import { APIDescription } from '../../apiDefs/schema';
-import { RootState } from '../../types';
-import AuthCodeFlowContentMDX from './AuthCodeFlowContent.mdx';
-
-const AuthCodeFlowContent = (): JSX.Element => {
-  const selectedOAuthApi = useSelector((state: RootState) => state.oAuthApiSelection.selectedOAuthApi);
-  const apiDef = lookupApiByFragment(selectedOAuthApi);
-  const selectorProps = {
-    options: getAllOauthApis().filter((item: APIDescription) => !isApiDeactivated(item)),
-    selectedOption: selectedOAuthApi,
-  };
-
-  const propObj = {
-    apiDef,
-    options: selectorProps.options,
-    selectedOption: selectorProps.selectedOption,
-  };
-
-  return (
-    <section aria-labelledby="authorization-code-flow">
-      <AuthCodeFlowContentMDX propObj={propObj} />
-    </section>
-  );
-};
-
-AuthCodeFlowContent.propTypes = {};
-
-export { AuthCodeFlowContent };
